+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * Validate Cart's total amount with the requested `total_amount`
   * Validation for `total_amount` (> 0)
 
+### Fixed
+
+  * Use production API URL by default
+
 
 ## [0.3.0] - 2023-04-03